Updated 3.4.5
3D Models now Accept tif Images - We added the tif extension to the allowed list of uploaded file types for 3D Models. Note that even though they upload, some of the images tested did not show until I converted them to jpg. It may be a browser or asynchronous timing when loading issue. If you test the 3D Model in a3D Scene, it appears to be consistent, either it loads correctly ever time or does not. If you see this issue, convert your images to jpg for better results.

Load All Quick Editor Setting - We have added a - Load All - Quick Editor Setting in Admin Mode. This allows you to load all Load Action Zones to show all sections of your map at once. While this is more memory intensive, it can be used to get snapshots and screen images showing any part of the 3D Scene at once. This only works in Admin Mode and has been implemented as a variable in the construct WTW.loadAllActionZones = 1. A button has been added into the Quick Editor Settings and in the Developer Tools menu. 

New 3D Compass - We are working replace the Browse Menu with the new HUD. We just replaced the old compass with a new 3D Compass that is in the 3D Scene. It attaches to the active camera so it will always be in front of the user camera. It will rotate to the correct angle of the avatar movement, zero degrees being North. In addition, it has a Red Arrow that will always point to the closest 3D Building in the 3D Scene (really helps when you wander off into the distance).

Toggle Compass - We replaced the toggleCompass function with hudToggleCompass in the wtw_hud.js file. This function loads the compass and parents it to the new camerafront transform node that stays 20 units in front of the current active camera. The new Babylon files, textures, and Blender files for the compass can be found in the /content/system/babylon/compass folder. You can press c on the keyboard to toggle the compass on-off.

Camera Front parent Box - We added a transform node called camerafront that is automatically added in front of the current active camera. This box becomes the attach point for anything you avatar would hold, compass, and the HUD. It uses billboard mode so that it always faces the camera.  This camerafront box is set to 20 units in front of the active camera by default.  You can use it as a parent box for any item, then set your offset as position, rotation, and scaling as needed.

Retired the UI Canvas - We are finally retiring the wtw_uicanvas from the main elements. This canvas was an overlay of the Babylon render Canvas that was used to add active elements in front of the render canvas. The last thing we used it for was the compass, which is now replaced by a 3D Compass with the HUD.

Moved Compass Movement to Render Cycle - We retired the Load UI Canvass function therefore moved the compass rose movement to the scene render loop cycle. This eliminated a call to the function and integrated to the sky dome movement for the avatar.

Aligned Sunrise to East on Compass - The past compass was 90 degrees out of phase with the sunrise ever since I changed the avatar rotation to match the 3D Scene rotation months ago. With the new compass in place, I swapped the settings under the quick sky settings to match Sunrise in the East and sunset in the West.  

Avatar Designer Loading - The Avatar designer was stuck on the loading screen. Ever since we added the Transform Nodes as the base of some 3D Objects (to speed up loading and rendering), we created a new function getMeshOrNodeByID as a global way to get an object regardless if it was a Mesh (mold) or Transform Node. It works great, but we forgot to add it to the Avatar Designer in the Avatar Plugin which has its own independent light-weight set of utilities that load with the designer. We added the function to that wtwavatars_common.js file and now the designer works as expected.

Avatar Enter Animations - We have reworked the Avatar Enter Animations! They were disabled when we rewrote the avatar designer a few months back. But now we have fixed the issues, reset some of the timing, and added the necessary clean-up elements. We reimplemented them as a way to enter a 3D Scene. Currently there are 11 animations and they will be chosen at random until we implement the new selection form in the new HUD.

Particle System Dispose Clean - The Particle systems were no longer being fully removed when we delete the parent object. Therefore we added code to the creation to tie the particle system variable to the object, then used it to check, stop, reset, and dispose of the particle system when it is called for disposeClean. 

Camera Selection moved to HUD - The HUD is accessible by h or m on the keyboard. Under settings, you can now select cameras. These settings will be replacing the browse menu camera selection. We have expanded the selections, changed the format, removed the function to loadCameraSettings and switchCamera, added initCamera and will be using loadPrimaryCamera to retrieve any other needed settings. The settings on the HUD will continue to evolve as we expand the functionality.

Init Camera Function - We created a simple initCamera function with a switch condition to create different cameras and specific settings. This replaces the individual functions for each init camera (initAnaglyphCamera, initVRCamera, initVRGamepadCamera, initFollowCamera, initArcCamera, and the function to call and add more active cameras, addActiveCamera). This should make it simper to add and implement new camera definitions.