Updated 3.4.5
Avatar Designer Loading - The Avatar designer was stuck on the loading screen. Ever since we added the Transform Nodes as the base of some 3D Objects (to speed up loading and rendering), we created a new function getMeshOrNodeByID as a global way to get an object regardless if it was a Mesh (mold) or Transform Node. It works great, but we forgot to add it to the Avatar Designer in the Avatar Plugin which has its own independent light-weight set of utilities that load with the designer. We added the function to that wtwavatars_common.js file and now the designer works as expected.