Updated 3.4.5
Enabled Reflections - In scenes with water, we once again enabled the use of reflections and refractions. The related functions now use the latest Water Material. We modified the add function and created a remove function. We also added the function calls to the babylonfile 3D Models mold type, Avatars, and disposeClean functions.