Updated 3.4.5
New 3D Compass - We are working replace the Browse Menu with the new HUD. We just replaced the old compass with a new 3D Compass that is in the 3D Scene. It attaches to the active camera so it will always be in front of the user camera. It will rotate to the correct angle of the avatar movement, zero degrees being North. In addition, it has a Red Arrow that will always point to the closest 3D Building in the 3D Scene (really helps when you wander off into the distance).

Fixed the Video Player - We completely reworked the video player! First we cleaned up the design in Blender and added node based coloring, fixed the animation on the buttons, orientation for new video players to be facing you, changed the scaling to be simple values (before it wanted you to keep it to the right scale ratio for the screen shape), hooked up the screen play and pause, fixed the video poster, and added the Transformation Node based Scaling Guides. 

Camera Selection moved to HUD - The HUD is accessible by h or m on the keyboard. Under settings, you can now select cameras. These settings will be replacing the browse menu camera selection. We have expanded the selections, changed the format, removed the function to loadCameraSettings and switchCamera, added initCamera and will be using loadPrimaryCamera to retrieve any other needed settings. The settings on the HUD will continue to evolve as we expand the functionality.