Updated 3.4.6
Version added to Connect Files - The version, versionid, and versiondesc are now part of the responses from connect files for Avatars, Buildings, Communities, and Things. This will assist with checking and controlling version updates.

Rename Version ID - Since the use of versionid is all the same for avatars, communities, buildings, and things, we decided to rename versionavatarid, versioncommunityid, versionbuildingid, and versionthingid all to versionid. This will also make it easier to compare versions and look for updates.

Added Version Description to Avatars - The Version Description has been added to the Share 3D Avatar form. This allows you to add up to 255 characters to describe what the update will include. This description will be seen when the users have the option to update to that version.