Updated 3.4.6
First Person Camera - The first person cameras for the normal universal camera are now parented properly and rotated to the right direction. The stable camera setting uses the standard follow camera moved to the avatar location parented to the avatar camera invisible box, which is used as a lock target when the camera is in follow mode. The motion first person camera uses the headtop as the parent which moves with the avatar animations. The first person camera is then offset to be 2 units forward so that it does not show any inside meshes like the avatar teeth! 

Added Avatar Animation Count - When you add or edit an Avatar, you have the option to add as many animations as you like to the avatar. After I counted them 3 or 4 times when I was trying to describe a new avatar I decided there is a better way. I now added a counted to the bottom of the list of animations loaded on the avatar. 

Alphabetized the Avatar Animations - We alphabetized the array of files returned from the getUploadedFileFileDetails PHP function. This allows us to have a sorted list of file names on file lists like the avatar animations drop down list. 

Auto Select Latest File - When adding avatar animations, when you upload a new file, the process will now set the latest file uploaded as the selected item in the dropdown list. It looks specifically for babylon files and will not select the babylon manifest. If more than one gets uploaded, it will select the last babylon file.

Avatar Camera Off Movement Speed - We increased the accuracy of the scroll wheel movement for the Avatar Camera Off used in Building under Admin Mode. Before you could not get to smaller incremental zoom levels. This is a balance in speed of zooming and increment of zoom. 

Default Animation File Set - In the Avatar Animations edit form, after the upload process was changed to set the last uploaded babylon file in the dropdown when added, the default for each animation edited was also changing to the new uploaded file. This has been corrected and both processes are setting the right file.

Opening Scene when Not Logged In - Last month we updated the cameras and all of the functions to the latest options available. When you opened the 3D Scene, before logging in, the camera was pointing straight up from the center location; always showing the sky until you logged in. Now the camera position is being activated even when you are not logged in and will show the scene in the camera view.

Updated Search Categories Samples - Since the Search Categories on the Share 3D Avatar form were copied from the 3D Things, the examples did not match what an avatar description would be like. We updated the category samples with Avatar, Leather Armor, and Boots as examples to describe the avatar with possible search terms.

Added Version Info to Copy Avatar - In the copy avatar process, the versionavatarid, version, and versiondesc have been added for the avatars table to be preserved when you copy a 3D Avatar.

Added Version Info to Save New Avatar - When saving a new 3D Avatar (Add New 3D Avatar or Create 3D Avatar), the version data was not set. It is now defaulting to the avatar id for versionavatarid, 1.0.0 for version, and Initial Version for versiondesc. This will make it easier to proceed with the next versions and upgrades.

Version added to Connect Files - The version, versionid, and versiondesc are now part of the responses from connect files for Avatars, Buildings, Communities, and Things. This will assist with checking and controlling version updates.

Rename Version ID - Since the use of versionid is all the same for avatars, communities, buildings, and things, we decided to rename versionavatarid, versioncommunityid, versionbuildingid, and versionthingid all to versionid. This will also make it easier to compare versions and look for updates.

Populate the Version ID in Tables - Since we changed to the versionid name, we needed to repopulate the versionid fields in the avatars, communities, buildings, and things tables. We also added the versiondesc with the words Initial Version for objects of version 1.0.0 

Share Avatar Version Form Section - On the Share 3D Avatar Form, there is a new section for version. It includes Initial Share, Update Share, Version, and Instructions. Only the Update Share option allows you to change the version number.

Version is now saved with 3D Avatar - The Version is now saved with the SaveAvatarTemplate PHP function. It will grab the value from the Share 3D Avatar Form.

Added Version Description to Avatars - The Version Description has been added to the Share 3D Avatar form. This allows you to add up to 255 characters to describe what the update will include. This description will be seen when the users have the option to update to that version. 

Avatar Version Added to Info Form - When editing an Avatar in Admin Mode, the 3D Avatar Information Form now displays the Version you are currently editing. This is not an editable field. You must use Share 3D Avatar option to create a new Version.

Select 3D Avatar to Edit Version - In Admin Mode, when you see the list to Select 3D Avatar to edit, it now shows the version (v#.#.#) number in parenthesis so that you can easily select from multiple versions of avatars with the same name.

Select 3D Community. Building, and Thing Version - In Admin Mode, the version has been added to all of the select 3D Web lists (3D Community, 3D Building, 3D Thing, and 3D Avatar). In addition, the version numbering has been aligned on the right side. 

Version added to 3D Community Information - When editing a 3D Community in Admin Mode, the 3D Community Information Form now displays the Version you are currently editing. This is not an editable field. You must use Share 3D Community option to create a new Version.

Version added to 3D Building Information - When editing a 3D Building in Admin Mode, the 3D Building Information Form now displays the Version you are currently editing. This is not an editable field. You must use Share 3D Building option to create a new Version.

Version added to 3D Thing Information - When editing a 3D Thing in Admin Mode, the 3D Thing Information Form now displays the Version you are currently editing. This is not an editable field. You must use Share 3D Thing option to create a new Version.