Updated 3.4.8
Global Role Access - When opening a 3D Community, 3D Building, or 3D Thing in the Admin Mode, the list was only returning the 3D Webs that you have granular access to edit. It now checks your logged in user for roles and will provide the complete list of 3D Webs available if you have the role of Admin, Developer, Architect, or Graphics Artist. This change applies to the connect files for communities, buildings, and things.