Updated 3.4.8
Image Plane and Image Uploads - When using a Plane and adding a texture, the texture was not reappearing when the webpage reloaded. It turns out that the image zwebimageid was blank. A test condition of setting zwebimageid to blank was not removed from a previous release testing. Since only one record can have a blank key field in the database, all additional image uploads were failing for duplicate key.  The line was removed from code and now all image uploads work again as expected.