Updated 3.4.8
Image Scaling on Planes - The image scaling (uscale and vscale) for planes was using a proportional formula that defaulted to tiling after a reload instead of scaling one image. The formula has been removed for a straight forward setting. 1x1 will be the original one image stretched to the plane size. A 3x3 will show 3 images vertically and 3 images horizontally. etc... This will make it easier to predict and set the desired behavior.