Updated 3.4.8
Display Name Input Value Populated - While the display name was being populated to the menu and other places, the main wtw_tdisplayname hidden input value was not being updated by some of the login functions (has various methods). We added code to the functions and also had it clear the value on all of the logout functions.