Updated 3.4.8
Edit User Roles - We have now added a way to edit user roles and add new roles. This will allow us to create groups of users... and in the near future set the permissions for what a user in a particular role can view. This means user groups that have access to certain things!

Add New User Role - Under the User Roles in Admin mode, you can now add a new Role.

Edit and Delete a Role - Under User Roles in Admin Mode, you can also edit an existing role name or delete a role. Using this form, you will not be able to delete some of the core roles like Admin, Architect, Developer, Graphics Artist, Subscriber, Guest or Host. 

Global Role Access - When opening a 3D Community, 3D Building, or 3D Thing in the Admin Mode, the list was only returning the 3D Webs that you have granular access to edit. It now checks your logged in user for roles and will provide the complete list of 3D Webs available if you have the role of Admin, Developer, Architect, or Graphics Artist. This change applies to the connect files for communities, buildings, and things.

Image Plane and Image Uploads - When using a Plane and adding a texture, the texture was not reappearing when the webpage reloaded. It turns out that the image zwebimageid was blank. A test condition of setting zwebimageid to blank was not removed from a previous release testing. Since only one record can have a blank key field in the database, all additional image uploads were failing for duplicate key.  The line was removed from code and now all image uploads work again as expected.

Image Scaling on Planes - The image scaling (uscale and vscale) for planes was using a proportional formula that defaulted to tiling after a reload instead of scaling one image. The formula has been removed for a straight forward setting. 1x1 will be the original one image stretched to the plane size. A 3x3 will show 3 images vertically and 3 images horizontally. etc... This will make it easier to predict and set the desired behavior.

Added Keystroke Hooks for Plugins - We added two new keystroke hooks so that plugins can intercept and code for Key Down and Key Up events. This will make it easier to customize keyboard mapping for movement or other triggers.

Display Name Input Value Populated - While the display name was being populated to the menu and other places, the main wtw_tdisplayname hidden input value was not being updated by some of the login functions (has various methods). We added code to the functions and also had it clear the value on all of the logout functions. 

Clear Select Mold Duplicate Call - In the Touch Down input control function, the WTW.clearSelectedMold function was being called 2 times back to back. This was an unnecessary duplicate and has been removed.

Plugins Onclick Expanded Conditions - Previously, the pluginOnClick event would only fire if the picked name of an object was not blank. We moved the function call to be before the picked name is checked so that plugins can use a non picked click (example: click on the sky or ground to clear a previously selected 3D object in the scene) to add their own code as needed to reset conditions, clear items, or execute a custom function when nothing is selected. 

Chat RoomID Expanded - In preparation for the addition of building based chats and multiplayer, we are expanding the ids for group chats and multiplayer (scene based) to include passing the communityid, buildingid, and thingid. The concept is that users in a 3D Community Scene can group chat and track multiplayer movement. If the owner of a 3D Building (for example 3D Shopping Store or Game) can open the 3D Building directly and chat with anyone from any 3D Community Scene that walks near one of the instances of the 3D Building. 

Added Plugins Add Action Zone Hook - We added a hook for when a load action zone is added. This allows us to tap into the one moment when a load action zone is added (executes once per load action zone) and avoid the meInActionZone that gets fired every cycle that the avatar remains in the load action zone.

data variable renamed to zdata - In the 3D Internet plugin for multiplayer and chat, the passed variable for multiplayer services was using the variable data. To match the existing naming standards that has been changed to zdata since it is an internal variable in a function. This includes the chat.js, move.js, and voicechat.js files. 

New Group Chat Text Prompt - There is a new Group Chat text Prompt that can be opened by hitting Return (while focused on the scene or mouse over scene). You can also hit return with the prompt blank to close it. If there is text, it will be set to the screen. We are in the early stages of passing the text to other users and requires more testing.

Text Block for Group Chat - When you submit text to the text prompt, the text is displayed on the screen (in scene, so that 3D VR can view it in scene). Each line will stay for 60 seconds then disappear. 

Enter and Exit Chat Functions - We have added an Enter Chat function that executes when your avatar enters a High - Load when far Load Zone and an Exit Chat function that executes when the avatar leaves the High - Load when far Load Zone. In the near future these functions will connect your group chat text to one or multiple chats based on the 3D Community scene you are in and any near 3D Building you are approaching or are in. It is like walking in and out of chat zones.