Updated 3.4.9
Doc Type on Pages Set - The doctype on certain auxiliary pages was not set and caused a warning for Google Chrome that may set the page in quirks mode. After setting the doctype to html, it forces the page to stay out of quirks mode. This applies to the loading, viewers, login, and other pages in the /core/pages folder.

Addition of Language Translation Files - We added language translation files to the /core/languages folder to define the word translations for the menus and forms built in WalkTheWeb. Additional files can be added and it will automatically load and use any properly formatted file in the folder. The language for the menu can be set in the admin menu - settings - server settings.

Admin Menu Supports Multiple Languages - The Admin Menus now support multiple languages. Just add a language file to /core/languages based on the format in the Spanish Example file. Then you can select the language for the menu in the Admin Menu - Settings - Server Settings.

PHP Language Translate Function - We added a PHP function called with 2 underscores __("translate words") that will return the matching word's translation from the language files. The translation is not case sensitive. The function is part of the wtwdb class and extends to be called from the wtwhandlers, wtwconnect, and wtwadminmenu classes.

JavaScript and PHP Translation Functions Added - Created a JavaScript function WTW.('word') and ('word') in PHP classes wtwdb, wtwhandlers, wtwconnect, and wtwtools that will translate the word based on the default language set and the translation file found in /core/languages. This should make it easier to adapt the translated words into WalkTheWeb menus and forms.