Updated 3.4.9
Enter and Exit Action Zone Handlers - We have added Enter and Exit Action Zone Handlers for plugins to use. Each function will fire one per Action Zone the avatar enters and fire once on exit. Note that when the browser looses focus it may trigger an Exit, and Enter again when focus is restored. 

Doc Type on Pages Set - The doctype on certain auxiliary pages was not set and caused a warning for Google Chrome that may set the page in quirks mode. After setting the doctype to html, it forces the page to stay out of quirks mode. This applies to the loading, viewers, login, and other pages in the /core/pages folder.

Addition of Language Translation Files - We added language translation files to the /core/languages folder to define the word translations for the menus and forms built in WalkTheWeb. Additional files can be added and it will automatically load and use any properly formatted file in the folder. The language for the menu can be set in the admin menu - settings - server settings.

Admin Menu Supports Multiple Languages - The Admin Menus now support multiple languages. Just add a language file to /core/languages based on the format in the Spanish Example file. Then you can select the language for the menu in the Admin Menu - Settings - Server Settings.

PHP Language Translate Function - We added a PHP function called with 2 underscores __("translate words") that will return the matching word's translation from the language files. The translation is not case sensitive. The function is part of the wtwdb class and extends to be called from the wtwhandlers, wtwconnect, and wtwadminmenu classes.

JavaScript and PHP Translation Functions Added - Created a JavaScript function WTW.('word') and ('word') in PHP classes wtwdb, wtwhandlers, wtwconnect, and wtwtools that will translate the word based on the default language set and the translation file found in /core/languages. This should make it easier to adapt the translated words into WalkTheWeb menus and forms.

Added Language Translation to 3D Plugins - Added the ability to add a translation file to a languages folder under the plugin root t translate any labels and terms used in the plugin. We also added the translation PHP function __() to the wtwplugins class. 

Server Settings Updates - In Admin Menu, Settings, Server Settings, we updated the layout to be wider to accommodate the labels especially for other languages. I also fixed the loading of the default language using the wtw_defaultlanguage PHP variable.

Corrected Settings Forms Issue - The forms under Admin Menu Settings were not opening correctly. There was a closing div tag missing causing the form sections to be skewed. 

Default Language Change Reload - On the Admin Menu - Settings - Server Settings page, when you change the default language for the menus, we added code to save the server settings form then reload the page with the new language showing. 

Multiplayer is now Zone Based - The multiplayer tracking used to be based on the 3D Community Scene. It is now based on the High Load Zones. The main benefit is that it only tracks and shows 3D Avatars in the same Load Zone as well as adding tracking by each 3D Building or 3D Thing separately. The additional tracking allows us to show 3D Avatars near a particular 3D Building when loaded directly, even when they are in different 3D Community Scenes. This makes it easier to work a virtual customer service job in a 3D Store that is located in more than one 3D Scene.

Alternative Language Spanish - The focus of the language translation is the menu system and the first alternative language included is Spanish. It starts with the json language file in the /core/languages folder and has the addition of a language file(s) in the plugins languages subfolders. This stage of release did not fully translate all of the menu forms or full page forms.