Updated 3.4.10
Fixed camera position. Recent camera updates appeared to have broken the dynamic camera distance function. This function allows the camera to automatically shorten the distance towards the avatar when the view is obstructed by a wall or mesh. It used to use radius of the free camera or universal camera, but that value no longer works. We now assigned that value and added code to use the value on the camera position formula in the setMovingCameras function.