Updated 3.4.11
Updated Copyright to include Database - In the browser console the Copyright statement is displayed. We added the database version, updated the date (which was missed on the last release), and reformatted the statement.

Unload Action Zone ID added to Molds Tables - Added the field for unloadactionzoneid as a varchar 16 to the communitymolds, buildingmolds, and thingmolds tables. This field will be used to track an unload zone or area within a load zone where the mold should not be loaded. 

Updated Unload Zone Functions - An Unload Zone is a Zone you can add inside of a Load Zone to force the opposite of a Load Zone. Where an avatar enters a Load Zone and the identified molds are loaded, the Unload Zone creates a hole in the Load Zone where the Molds are not loaded. Example would be a building with a lot of details outside loaded in a load zone, yet when you are inside the unload zone inside the building, the outside details can be unloaded to increase the available resources for the inside scene. This affects the following functions: wtw_common.js function setShownMoldsByWeb excludes the identified molds from the load zone from being loaded; wtw_basicactionzones.js function addActionzoneUnloadzone to create the zone; wtw_addactionzonelist.js to define the add function, default form values, and form labels; and wtw_analytics, 3dinternet class_main.js, and 3dinternet chat.js functions needed to exclude unloadzone from loadzone filter.

ztestactionzoneid eliminated in loadMolds function - The loadMolds function in wtw_core.js set a variable named ztestactionzoneid but never used the variable after it was set. We eliminated its use in the function declaration and setting. 

Exclusion Zone = Unload Zone on Mold Form - When editing a mold, under the Advanced Options section, you can now add an exclusion zone (unload zone) to the mold or 3D Model to have it unload while you are in a particular zone. This creates holes in the load zone allowing you to save on resources for molds you cannot see from the inside (like outside of a building not being shown when inside the building).  Note that the Exclusion Zone only appears if you have added an Unload Zone to your 3D Community, 3D Building, or 3D Thing being edited.

Connect, Share, Copy, and Recovery for Unload Zone - The unload zone id needed to be added to every instance of the Molds arrays, connect, save functions, copy functions, recover molds, share, download and all other references to mold architecture. This affected 151 instances in 24 files.