Updated 3.4.11
Added Version Support to Downloads - Added versionid, version, and versiondesc to the various handlers, connect, and class functions related to communities, buildings, things, and avatars. This stage included making sure the version information travels with the WalkTheWeb Downloads and Shared items. Furthermore it updated the functions for save, download, copy, and import of these items. 

New Group Chat Functionality! - We now have new group chat functionality! If you click Enter while the canvas is focused, it will open a text prompt. Just type and hit enter to send. The scope is currently the entire Community, Building, or Thing that is open. Your own posts show in yellow text, while all others show up in white. Hit the enter with a blank text top close the text prompt, or use the x on the prompt frame.

Updated Private Chat - The Private chat has been updated to include any avatar in your scene (community, building, or thing). Just click on the avatar, select private chat from the menu, and if they accept, you are off and running. If you close the chat, the other user gets the notification that you left. This chat also includes the update for "Typing" when the user is actively typing into the chat (before they send).