Updated 3.4.12
Added 3D Internet Blocked Instances Table - Added a table called wtwprefix_3dinternet_blockedinstances to track blocked and banned instances from your scenes on a user level. This functionality will be tied to multiplayer movement and chat functionality.

Looping Backwards through an Array Correction - Numerous times in the code we looped backwards through an array as noted by the i-- most often used to remove an index. By going backwards, when you remove the index, the next index is in tact to continue the process, unlike when you go forward and it shifts the remaining indexes while skipping one. We found a lack of consistency and sometimes an error in the method where it would skip an index. We searched and replaced the code to be more consistent with this pattern: for (var i=array.length-1; i > -1; i--) ... Note that some did not use the -1 from the length and just trapped the out of index by checking as null, while others used i >=0 which did work. But we found cases where it used i > 0 which does skip the lowest index. They are all corrected now.

Avatar Double loading in Multiplayer - Corrected an issue where avatars were loading twice while in multiplayer. They were downloading once, but playing the enter avatar animation twice. Corrected the issue by checking the second event to see if it already loaded.

Added JavaScript Hooks for Avatar Enter Events - We added two new hooks in the JavaScript code to make it easy to add custom code in plugins. The first is when the Avatar Enters the scene (part of the loading) and the second is when the avatar is completely loaded and visible. pluginsEnterAvatar and pluginsAvatarLoadComplete both pass the parameter zavatarname. 

Block Chat or Ban Avatar - You can now block chat or ban an avatar by clicking on the avatar in the 3D Scene and selecting it from the menu. When you Block Chat, the avatar in your scene and your avatar in their scene will change to 50% transparency, you will not be able to start a private chat, you will not see any group chat from them, and they will not get see any group chat from you. If you Ban the User, the avatar will go to full transparency and all of the Block Chat functions are also in effect.