Updated 3.4.12
Rebranded the Styled the Install Process - We rebranded and styled the Install Process so that it uses the same logo that displays on the bottom left, black background based, and larger fonts. We also moved inline styles to the stylesheet. 

New Install Step - Optional Services (Multiplayer) - We now added an additional screen to the setup process to offer Optional Services. The first service offered is to enable multiplayer. There are 3 basic multiplayer plans to start; Developer, Small Business, and Gamers. You can either purchase a year of service or simply click No Thanks. After the step is shown once, it will not show again. The same settings and ability to turn on Multiplayer is available in the Admin Menu, 3D Internet, Control Panel.