Updated 3.4.12
Updated Socket.io to the Latest Release - We updated socket.io to version 4.4.1 (latest release). This is for the multiplayer  and chat functionality.

Renamed Plugin Hook pluginsCheckActionPerZone - We will be adding a new hook in the check action zone function that is not in the zone array loop so we renamed the existing plugin function from pluginsCheckActionZone to pluginsCheckActionPerZone since it iterates through the array of zones.. We also renamed pluginsCheckActionZoneTrigger to pluginsCheckActionPerZoneTrigger for consistency since it is also in the iteration of the array.

Added Mold ID to 3D Text - In the addMold3DText function on script page wtw_3dblog.js under molds, when the text is added it used a mold name but was missing the ID. Without the id, it could not be selected with getMeshById functions. We added the ID to the mesh to match the name to resolve this issue.

New JavaScript Hook for Reset Activity Timer - We added a new hook for when my avatar movement resets the activity timer. This timer is currently being used to pause the scene rendering if the avatar is not moved within a set amount of time. For mobile devices it is relatively short while on desktop devices it is extremely long. Reason has to do with battery saving concerns on mobile devices. This new hook allows plugins to tap into the process and set the reset loop differently or add functionality to when it is reset. 

Remove getMoveEventsArray Function - We removed the unused getMoveEventsArray JavaScript Function from the 3D Internet Plugin.

Enabled Activity Timer - The Activity Timer has been disabled for a while since it had no extended use. With Multiplayer in full swing, we now needed the activity timer to decide when the inactivity timeout occurs so that we can gracefully disconnect form multiplayer, leading to a more reliable reconnect sequence. The Activity Timer is activated with the function resetActivityTimer, initially called in the loadUserSettingsAfterEngine function found in the wtw_core.js file.

Enhanced Multiplayer Functionality - We improved multiplayer functionality with special concentration on making sure avatars leave when disconnected and return as expected. This includes a timeout for inactivity that after 30 minutes your avatar will fade in your 3D Scene and in the 3D Scenes you are shown in multiplayer, then you have 5 minutes to either move you avatar to reset the timer or it will log you out of multiplayer until you do move to reenable it.

Added inactive Flag for Multiplayer Movement - In order to stop all multiplayer sending and receiving commands, we added a global variable wtw3dinternet.inactive Boolean flag that can be used to test if the activity timer timed out and multiplayer is temporary on hold. 

Reconnect Load Zone Function - When a user disconnects all socket connections and rooms disconnect, so when we reconnect we needed a process to determine which Load Zones you are currently in and log you into the respective movement and group chat rooms.

Fade Multiplayer Avatar Listener - We added a Listener to the Movement (move channel object) to allow us to receive fade notifications caused by inactivity and returned activity. Fade can be used to obviously fade an avatar, but also can restore the opacity of an avatar using the same function. 

Add Avatar Parameter - We added a box around your avatar that can determine when other avatars in multiplayer are near you. When they are near, they will be shown and fade into view and when they leave the avatar parameter they will fade out. In addition, multiplayer avatars are now being tracked anywhere within the Extreme Load Zone. 

Multiplayer Script Documentation - The multiplayer 3D Internet Plugin now has a lot more comments and notes about each variable and function. This should help determine the use of each.