Updated 3.4.13
Added Downloads Queue Table - We added a new table called downloads that will queue up potential WalkTheWeb 3D Downloads to your server. At this time, it is for 3D Communities, 3D Buildings, 3D Things, and 3D Avatars. The intent is to populate this table when items are requested to be downloaded from other sources (like the walktheweb.org website), then pending an admin approval on your server, the items will complete the download and add to your available 3D Webs respectively. 

Added Download Date and UserID to Downloads Table - We added downloaddate and downloaduserid to downloads table to assist in determining if an item in the queue was downloaded and deleted or just canceled and deleted.

www.walktheweb.org Now Supports Downloads Search - We now added the WalkTheWeb Downloads functionality to the www.walktheweb.org website (work in progress to support the GitHub project and users/developers). These are the same downloads you can find within WalkTheWeb open-source. When you click download, you will be prompted for your WalkTheWeb URL and the system will add the download to a queue found on your Admin Dashboard. It will still require someone with admin permissions to download, but it makes it easier to prompt the download list.

Admin Dashboard Downloads Queue - On the WalkTheWeb open-source Admin Dashboard, when you have any downloads in the Queue (prompted from www.walktheweb.org website), they will show with the options to download or cancel. This downloads queue window only shows when there are downloads pending. 

Toggle Admin Dashboard Windows - The toggle arrow to maximize or minimize the Admin Dashboard windows was not working properly. Corrected the code to hide or show the contents of the box for the resizing.

Downloads Functions added to the Communities Class  - We added two functions to the wtw_communities.php class file to handle adding and updating records in the downloads queue. Add uses the addDownloadQueue function and supports the Connect class, while updateDownloadsQueue function is used by the handlers class to update a record to downloaded or cancelled in the downloads table.

Downloads Queue Handler for Updates - The Handler for the Communities Class now has the addition of updatedownloadqueue function case to process the updates to the downloads queue as downloaded or cancelled. Downloaded 3D Web records will have a downloaddate and downloaduserid in the table, both conditions mark the file deleted (processed and flagged to not show in the queue). No records are actually deleted.

Dashboard Connect Additions - The Admin Dashboard uses the connect/dashboard.php file to load the contents. We added code to include any downloads in the queue. If found, the process will use this information to validate the download and add additional information to it, such as preview image, title, description, version, and author.

Preview Feed for Downloads Queue - The downloads queue uses the minimum key field data to determine the item to download. In order for the Dashboard page to show the download information and preview image, we added a service on 3dnet.walktheweb.com that receives a set of downloadid, webid, and webtype (community, building, thing, or avatar) and returns the same data populated with the download information needed for the preview. The JavaScript function getDownloadsInfo in wtw_adminforms.js performs this process.