Updated 3.4.13
Corrected Image Previews for Avatars - When selecting an avatar, some of the legacy avatars used a different defined path. We updated the path to check for alternate path images before defaulting to a missing image type icon. 

Added Ability to Delete My Avatar from List - When logging in, you are given a chance to select your avatar from a list of your previously used avatars. If you scroll to the bottom, we added a button to Edit My Avatars List. When selected it adds a delete button next to each avatar in your list. Select delete to flag the user avatar as deleted and no longer to show on the list. The actual records are not deleted. To cancel, click the Cancel - Edit My Avatars Button that replaces the original Edit button.

Updated pluginsGetMyAvatarList to support Edit Mode Parameter - We updated the existing pluginsGetMyAvatarList plugin hook to add a parameter determining if the My Avatars List is in Edit mode or not. zeditmode is passed and can be a Boolean true or false. If it is undefined, it defaults to false. 

Delete User Avatar Functionality Added - We added a process for flagging a user avatar as deleted, where it will not show up on any of the lists. The process includes a deleteUserAvatar function in core/scripts/avatars/wtw_avatarfunctions.js, deleteuseravatar case in the /core/handlers/avatars.php, and a deleteUserAvatar function in the /core/functions/class_wtwavatars.php class.  At this time, the avatar can be recovered only through setting the deleted column to zero in the useravatars database table.